Make patch_allowed_actions into patch_class_field
which can be used to patch other class fields for Resource tests. Also
move it to a utils.py, so that other tests can reuse it.
